+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2022-07-07
+
+- Added translations
+  - Added pluralization example
+  - Added formatter example (with Luxon)
+  - Used HTTP loader
+- Added a suspense fallback page for app loading
+- Added cypress eslint rules
+
 ## [0.2.0] - 2022-06-27
 
 - Added [cypress.io](https://www.cypress.io/)